Use dependency management to manage versions
Using dependency management we don't need to configure versions at each sub module. The submodules did make use of a variable, that is inherited from the root POM, so we might as well inherit the version from dependency management and avoid having to configure the version. Having the version in dependency management also makes the project compatible with maven's release plugin which would be able to automatically set the SNAPSHOT versions to proper versions